tests(MockCsvSensor): Add mock sensor to use pre-recorded data for testing
This allows future tests to use pre-recorded CSV sensor events for testing. Please
note that this is not a Sensor or SensorEvent as can be found in android.hardware as
these are considered final and do not allow to be extended or directly replaced in any way.
Therefore the small mock classes have been created for improved testability.